refactor: split monolithic tools into focused modules
- Extract resume/cover-letter text parsers to lib/resume_parser.py (881 lines)
  tools/export.py: 1,035 → 184 lines (PDF rendering + MCP tools only)

- Extract pipeline helpers to pipeline_helpers.py (654 lines)
  pipeline.py: 1,696 → 1,080 lines (route handlers + _pipeline_payload only)

- Extract generate constants to tools/generate_prompts.py (~270 lines)
  tools/generate.py: 1,618 → 1,351 lines

All 650 tests pass.
